Move toEmitAnything to own file
