+++ packages/apollo-angular/test-utils/matchers.ts
@@ -1,34 +1,6 @@
 import { expect } from 'vitest';
-import { ObservableStream, TakeOptions } from './ObservableStream';
+import { toEmitAnything } from './matchers/toEmitAnything';
 
 expect.extend({
-  async toEmitAnything(actual: ObservableStream<any>, options: TakeOptions) {
-    const stream = actual as ObservableStream<any>;
-    const hint = this.utils.matcherHint('toEmitAnything', 'stream');
-
-    try {
-      const value = await stream.peek(options);
-
-      return {
-        pass: true,
-        message: () => {
-          return (
-            hint +
-            '\n\nExpected stream not to emit anything but it did.' +
-            '\n\nReceived:\n' +
-            this.utils.printReceived(value)
-          );
-        },
-      };
-    } catch (error) {
-      if (error instanceof Error && error.message === 'Timeout waiting for next event') {
-        return {
-          pass: false,
-          message: () => hint + '\n\nExpected stream to emit an event but it did not.',
-        };
-      } else {
-        throw error;
-      }
-    }
-  },
+  toEmitAnything,
 });
